requirements.txt
customtkinter
Pillow
arabic-reshaper
python-bidi
matplotlib
SQLAlchemy 